+++ spec/clean-test-resources.sh
@@ -4,13 +4,14 @@ for fun in $functions; do
   aws lambda delete-function --function-name $fun
 done
 
-#apis=`aws apigateway get-rest-apis --query 'items[?starts_with(name,\`test\`)].id' --output text`
-#for api in $apis; do
-#  echo deleting API $api
-#  aws apigateway delete-rest-api --rest-api-id $api
-#done
+apis=`aws apigateway get-rest-apis --query 'items[?starts_with(name,\`test\`)].id' --output text`
+for api in $apis; do
+  echo deleting API $api
+  sleep 10 
+  aws apigateway delete-rest-api --rest-api-id $api
+done
 
-#roles=`aws iam list-roles --query 'Roles[?starts_with(RoleName, \`test\`)].RoleName' --output text`
+roles=`aws iam list-roles --query 'Roles[?starts_with(RoleName, \`test\`)].RoleName' --output text`
 
 
 lambdaLogs=`aws logs describe-log-groups --query 'logGroups[?starts_with(logGroupName,\`/aws/lambda/test\`)].logGroupName' --output text`
@@ -25,3 +26,9 @@ for log in $apiLogs; do
   aws logs delete-log-group --log-group-name $log
 done
 
+buckets=`aws s3api list-buckets --output text --query 'Buckets[?starts_with(Name,\`test\`)].Name'`
+for bucket in $buckets; do
+  echo deleting bucket $bucket
+  aws s3 rm --recursive s3://$bucket
+  aws s3 rb s3://$bucket
+done