+++ spec/test-projects/local-dependencies/local_modules/tst/index.js
@@ -1,0 +1,5 @@
+/*global module */
+module.exports = function () {
+	'use strict';
+	return 'hello local';
+};