+++ spec/test-projects/local-dependencies/main.js
@@ -1,0 +1,6 @@
+/*global require, exports */
+var localDep = require('tst');
+exports.handler = function (event, context) {
+	'use strict';
+	context.succeed(localDep());
+};