+++ src/tasks/rebuild-web-api.js
@@ -3,6 +3,8 @@ var aws = require('aws-sdk'),
 	Promise = require('bluebird'),
 	templateFile = require('../util/template-file'),
 	validHttpCode = require('../util/valid-http-code'),
+	validAuthType = require('../util/valid-auth-type'),
+	validCredentials = require('../util/valid-credentials'),
 	allowApiInvocation = require('./allow-api-invocation'),
 	pathSplitter = require('../util/path-splitter'),
 	promiseWrap = require('../util/promise-wrap'),
@@ -64,11 +66,12 @@ module.exports = function rebuildWebApi(functionName, functionVersion, restApiId
 				}
 			});
 		},
-		putLambdaIntegration = function (resourceId, methodName) {
+		putLambdaIntegration = function (resourceId, methodName, credentials) {
 			return apiGateway.putIntegrationAsync({
 				restApiId: restApiId,
 				resourceId: resourceId,
 				httpMethod: methodName,
+				credentials: credentials,
 				type: 'AWS',
 				integrationHttpMethod: 'POST',
 				requestTemplates: {
@@ -115,6 +118,25 @@ module.exports = function rebuildWebApi(functionName, functionVersion, restApiId
 				},
 				apiKeyRequired = function () {
 					return methodOptions && methodOptions.apiKeyRequired;
+				},
+				authorizationType = function () {
+					if (methodOptions && methodOptions.authorizationType && validAuthType(methodOptions.authorizationType.toUpperCase())) {
+						return methodOptions.authorizationType.toUpperCase();
+					} else if (methodOptions && (methodOptions.invokeWithCredentials === true || validCredentials(methodOptions.invokeWithCredentials))) {
+						return 'AWS_IAM';
+					} else {
+						return 'NONE';
+					}
+				},
+				credentials = function () {
+					if (methodOptions && methodOptions.invokeWithCredentials) {
+						if (methodOptions.invokeWithCredentials === true) {
+							return 'arn:aws:iam::*:user/*';
+						} else if (validCredentials(methodOptions.invokeWithCredentials)) {
+							return methodOptions.invokeWithCredentials;
+						}
+					}
+					return null;
 				},
 				isRedirect = function (code) {
 					return /3[0-9][0-9]/.test(code);
@@ -220,13 +242,13 @@ module.exports = function rebuildWebApi(functionName, functionVersion, restApiId
 					});
 				};
 			return apiGateway.putMethodAsync({
-				authorizationType: 'NONE', /*todo support config */
+				authorizationType: authorizationType(),
 				httpMethod: methodName,
 				resourceId: resourceId,
 				restApiId: restApiId,
 				apiKeyRequired: apiKeyRequired()
 			}).then(function () {
-				return putLambdaIntegration(resourceId, methodName);
+				return putLambdaIntegration(resourceId, methodName, credentials());
 			}).then(function () {
 				var results = [{code: successCode(), pattern: '', contentType: successContentType(), template: successTemplate(headers('success')), headers: headers('success')}];
 				if (errorCode() !== successCode()) {
@@ -238,7 +260,7 @@ module.exports = function rebuildWebApi(functionName, functionVersion, restApiId
 		},
 		createCorsHandler = function (resourceId, allowedMethods) {
 			return apiGateway.putMethodAsync({
-				authorizationType: 'NONE', /*todo support config */
+				authorizationType: 'NONE',
 				httpMethod: 'OPTIONS',
 				resourceId: resourceId,
 				restApiId: restApiId