+++ src/util/valid-credentials.js
@@ -1,0 +1,6 @@
+/*global module */
+module.exports = function validCredentials(creds) {
+	'use strict';
+	var credsRegex = /^arn:aws:(iam|sts)::(\*|\d{12})?:/;
+	return (typeof creds === 'string') && credsRegex.test(creds);
+};